+++ listen-engine/src/server/mod.rs
@@ -2,50 +2,17 @@ use actix_cors::Cors;
 use actix_web::{
     middleware,
     web::{self, Data},
-    App, HttpRequest, HttpResponse, HttpServer, Responder,
+    App, HttpResponse, HttpServer, Responder,
 };
 use std::sync::Arc;
 use tokio::sync::mpsc;
-use tokio::sync::oneshot;
-use uuid::Uuid;
 
-use crate::{
-    engine::{
-        api::{PipelineParams, WirePipeline},
-        error::EngineError,
-        pipeline::Pipeline,
-        Engine,
-    },
-    metrics::metrics_handler,
-};
+use crate::{engine::Engine, metrics::metrics_handler, server::state::AppState};
 use privy::{config::PrivyConfig, Privy};
 
-#[derive(Debug)]
-pub enum EngineMessage {
-    AddPipeline {
-        pipeline: Pipeline,
-        response_tx: oneshot::Sender<Result<String, EngineError>>,
-    },
-    GetPipeline {
-        user_id: String,
-        pipeline_id: Uuid,
-        response_tx: oneshot::Sender<Result<Pipeline, EngineError>>,
-    },
-    DeletePipeline {
-        user_id: String,
-        pipeline_id: Uuid,
-        response_tx: oneshot::Sender<Result<(), EngineError>>,
-    },
-    GetAllPipelinesByUser {
-        user_id: String,
-        response_tx: oneshot::Sender<Result<Vec<Pipeline>, EngineError>>,
-    },
-}
-
-pub struct AppState {
-    engine_bridge_tx: mpsc::Sender<EngineMessage>,
-    privy: Arc<Privy>,
-}
+pub mod create;
+pub mod get;
+pub mod state;
 
 pub async fn run() -> std::io::Result<()> {
     let (server_tx, server_rx) = mpsc::channel(1000);
@@ -99,8 +66,8 @@ pub async fn run() -> std::io::Result<()> {
             )
             .wrap(middleware::Logger::default())
             .route("/healthz", web::get().to(healthz))
-            .route("/pipeline", web::post().to(create_pipeline))
-            .route("/pipelines", web::get().to(get_pipelines))
+            .route("/pipeline", web::post().to(create::create_pipeline))
+            .route("/pipelines", web::get().to(get::get_pipelines))
             .route("/metrics", web::get().to(metrics_handler))
     })
     .bind(("0.0.0.0", 6966))?
@@ -135,191 +102,4 @@ async fn healthz() -> impl Responder {
     HttpResponse::Ok().json(serde_json::json!({
         "status": "healthy"
     }))
-}
-
-async fn get_pipelines(state: Data<AppState>, req: HttpRequest) -> impl Responder {
-    let auth_token = match req.headers().get("authorization") {
-        Some(auth_token) => auth_token.to_str().unwrap(),
-        None => {
-            return HttpResponse::Unauthorized().json(serde_json::json!({
-                "status": "error",
-                "message": "Authorization header is required"
-            }));
-        }
-    };
-    let auth_token = auth_token.split(" ").nth(1).unwrap();
-
-    let user = match state
-        .privy
-        .authenticate_user(auth_token)
-        .await
-        .map_err(|_| HttpResponse::Unauthorized())
-    {
-        Ok(user) => user,
-        Err(_) => {
-            return HttpResponse::Unauthorized().json(serde_json::json!({
-                "status": "error",
-                "message": "Unauthorized"
-            }));
-        }
-    };
-
-    let (response_tx, response_rx) = oneshot::channel();
-    tracing::debug!("Sending GetAllPipelinesByUser message to engine");
-    match state
-        .engine_bridge_tx
-        .send(EngineMessage::GetAllPipelinesByUser {
-            user_id: user.user_id.clone(),
-            response_tx,
-        })
-        .await
-    {
-        Ok(_) => tracing::debug!("Successfully sent message to engine"),
-        Err(e) => {
-            tracing::error!("Failed to send message to engine: {}", e);
-            return HttpResponse::InternalServerError().json(serde_json::json!({
-                "status": "error",
-                "message": "Engine communication error"
-            }));
-        }
-    };
-
-    tracing::debug!("Waiting for response from engine");
-    let pipelines = match response_rx.await {
-        Ok(Ok(pipelines)) => {
-            tracing::debug!("Received {} pipelines from engine", pipelines.len());
-            pipelines
-        }
-        Ok(Err(e)) => {
-            tracing::error!("Engine error: {}", e);
-            return HttpResponse::InternalServerError().json(serde_json::json!({
-                "status": "error",
-                "message": format!("Failed to get pipelines: {}", e)
-            }));
-        }
-        Err(e) => {
-            tracing::error!("Channel closed: {}", e);
-            return HttpResponse::InternalServerError().json(serde_json::json!({
-                "status": "error",
-                "message": "Internal communication error"
-            }));
-        }
-    };
-
-    HttpResponse::Ok().json(serde_json::json!({
-        "status": "success",
-        "pipelines": pipelines
-    }))
-}
-
-async fn create_pipeline(
-    state: Data<AppState>,
-    req: HttpRequest,
-    wire: web::Json<WirePipeline>,
-) -> impl Responder {
-    let start = std::time::Instant::now();
-
-    let auth_token = match req.headers().get("authorization") {
-        Some(auth_token) => auth_token.to_str().unwrap(),
-        None => {
-            return HttpResponse::Unauthorized().json(serde_json::json!({
-                "status": "error",
-                "message": "Authorization header is required"
-            }));
-        }
-    };
-    let auth_token = auth_token.split(" ").nth(1).unwrap();
-
-    let user = match state
-        .privy
-        .authenticate_user(auth_token)
-        .await
-        .map_err(|_| HttpResponse::Unauthorized())
-    {
-        Ok(user) => user,
-        Err(_) => {
-            return HttpResponse::Unauthorized().json(serde_json::json!({
-                "status": "error",
-                "message": "Unauthorized"
-            }));
-        }
-    };
-
-    if user.pubkey.is_none() {
-        return HttpResponse::Unauthorized().json(serde_json::json!({
-            "status": "error",
-            "message": "Wallet is required in order to create a pipeline"
-        }));
-    }
-
-    metrics::counter!("pipeline_creation_attempts", 1);
-
-    let pipeline: Pipeline = (
-        wire.into_inner(),
-        PipelineParams {
-            user_id: user.user_id,
-            wallet_address: user.wallet_address.clone(),
-            pubkey: user.pubkey.clone(),
-        },
-    )
-        .into();
-
-    tracing::info!(pipeline = ?pipeline, "creating pipeline");
-
-    // Create oneshot channel for response
-    let (response_tx, response_rx) = oneshot::channel();
-
-    // Send message to engine
-    if let Err(e) = state
-        .engine_bridge_tx
-        .send(EngineMessage::AddPipeline {
-            pipeline,
-            response_tx,
-        })
-        .await
-    {
-        metrics::counter!("pipeline_creation_errors", 1);
-        return HttpResponse::InternalServerError().json(serde_json::json!({
-            "status": "error",
-            "message": format!("Failed to communicate with engine: {}", e)
-        }));
-    }
-
-    // Wait for response with timeout
-    let result = match tokio::time::timeout(std::time::Duration::from_secs(5), response_rx).await {
-        Ok(response) => match response {
-            Ok(Ok(id)) => {
-                metrics::counter!("pipeline_creation_success", 1);
-                HttpResponse::Created().json(serde_json::json!({
-                    "status": "success",
-                    "message": "Pipeline created successfully",
-                    "id": id
-                }))
-            }
-            Ok(Err(e)) => {
-                metrics::counter!("pipeline_creation_errors", 1);
-                HttpResponse::InternalServerError().json(serde_json::json!({
-                    "status": "error",
-                    "message": format!("Failed to create pipeline: {}", e)
-                }))
-            }
-            Err(e) => {
-                metrics::counter!("pipeline_creation_errors", 1);
-                HttpResponse::InternalServerError().json(serde_json::json!({
-                    "status": "error",
-                    "message": format!("Failed to receive response from engine: {}", e)
-                }))
-            }
-        },
-        Err(_) => {
-            metrics::counter!("pipeline_creation_errors", 1);
-            HttpResponse::GatewayTimeout().json(serde_json::json!({
-                "status": "error",
-                "message": "Pipeline creation timed out"
-            }))
-        }
-    };
-
-    metrics::histogram!("pipeline_creation_duration", start.elapsed());
-    result
 }
